Changin files and folder names to match javascript naming convetion.
