+++ source/delivery-strategy/containerizedDeliveryStrategy.ts
@@ -1,12 +1,20 @@
 import { ResolutionContext } from "../ResolutionContext";
 import { DependencyMetadata } from "../DependencyMetadata";
 import { IDeliveryStrategy } from "./IDeliveryStrategy";
-import { DeliveryError } from "../Errors/DeliveryError"
+import { DeliveryError } from "../errors/DeliveryError"
+import { IContainer } from "../IContainer";
 
 /**
- * Represenst an strategy to deliver a new instance targets with an specific strategy.
+ * Represenst an strategy to deliver a new instance targets per container with an specific strategy.
  */
-export class PerCallDeliveryStrategy implements IDeliveryStrategy {
+export class ContainerizedDeliveryStrategy implements IDeliveryStrategy {
+
+    private _containerInstanceMap:{container:IContainer, instance:any}[];
+
+    constructor() {
+        this._containerInstanceMap = [];
+    }
+
     /**
      * Deliver the transformed reference in the provided dependency metadata.
      * @param resolutionContext Represents the context in which the request was made.
@@ -17,6 +25,9 @@ export class PerCallDeliveryStrategy implements IDeliveryStrategy {
         if (!resolutionContext)
             throw new DeliveryError('Must provide a valid resolution context.');
 
+        if (!resolutionContext.originContainer)
+            throw new DeliveryError('The provided resolution context must have a valid origin container.');
+
         if (!dependencyMetadata)
             throw new DeliveryError('Must provide the depencency metadata to deliver from.');
 
@@ -26,6 +37,16 @@ export class PerCallDeliveryStrategy implements IDeliveryStrategy {
         if (!dependencyMetadata.servicingStrategy)
             throw new DeliveryError('The provided dependency metadata must have a valid servicing strategy.')
 
-        return dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);        
+        let map = this._containerInstanceMap.find(map => map.container === resolutionContext.originContainer);
+        let servingResult:any;
+        
+        if (!map) {
+            servingResult = dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);
+            this._containerInstanceMap.push({ container: resolutionContext.originContainer, instance: servingResult });
+        } else
+            servingResult = map.instance;
+
+
+        return servingResult;        
     }
 }
