+++ source/delivery-strategy/iDeliveryStrategy.ts
