+++ source/delivery-strategy/perCallDeliveryStrategy.ts
@@ -1,19 +1,12 @@
 import { ResolutionContext } from "../ResolutionContext";
 import { DependencyMetadata } from "../DependencyMetadata";
 import { IDeliveryStrategy } from "./IDeliveryStrategy";
-import { DeliveryError } from "../Errors/DeliveryError"
+import { DeliveryError } from "../errors/DeliveryError"
 
 /**
- * Represenst an strategy to deliver a new instance targets per resolution with an specific strategy.
+ * Represenst an strategy to deliver a new instance targets with an specific strategy.
  */
-export class PerResolutionDeliveryStrategy implements IDeliveryStrategy {
-
-    private _resolutionContextInstanceMap:{context:ResolutionContext, instance:any}[];
-
-    constructor() {
-        this._resolutionContextInstanceMap = [];
-    }
-
+export class PerCallDeliveryStrategy implements IDeliveryStrategy {
     /**
      * Deliver the transformed reference in the provided dependency metadata.
      * @param resolutionContext Represents the context in which the request was made.
@@ -33,16 +26,6 @@ export class PerResolutionDeliveryStrategy implements IDeliveryStrategy {
         if (!dependencyMetadata.servicingStrategy)
             throw new DeliveryError('The provided dependency metadata must have a valid servicing strategy.')
 
-        let map = this._resolutionContextInstanceMap.find(map => map.context === resolutionContext);
-        let servingResult:any;
-        
-        if (!map) {
-            servingResult = dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);
-            this._resolutionContextInstanceMap.push({ context: resolutionContext, instance: servingResult });
-        } else
-            servingResult = map.instance;
-
-
-        return servingResult;        
+        return dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);        
     }
 }
