+++ source/delivery-strategy/singletonDeliveryStrategy.ts
@@ -1,19 +1,22 @@
 import { ResolutionContext } from "../ResolutionContext";
 import { DependencyMetadata } from "../DependencyMetadata";
 import { IDeliveryStrategy } from "./IDeliveryStrategy";
-import { DeliveryError } from "../Errors/DeliveryError"
-import { IContainer } from "../IContainer";
+import { DeliveryError } from "../errors/DeliveryError"
 
 /**
- * Represenst an strategy to deliver a new instance targets per container with an specific strategy.
+ * Represenst an strategy to deliver the same instance targets with an specific strategy.
  */
-export class ContainerizedDeliveryStrategy implements IDeliveryStrategy {
+export class SingletonDeliveryStrategy implements IDeliveryStrategy {
 
-    private _containerInstanceMap:{container:IContainer, instance:any}[];
+    /**
+     * Represents a boolean value specifying if the target has been served.
+     */
+    private _isServed:boolean;
 
-    constructor() {
-        this._containerInstanceMap = [];
-    }
+    /**
+     * Represets an instance of the target.
+     */
+    private _instance:any;
 
     /**
      * Deliver the transformed reference in the provided dependency metadata.
@@ -25,9 +28,6 @@ export class ContainerizedDeliveryStrategy implements IDeliveryStrategy {
         if (!resolutionContext)
             throw new DeliveryError('Must provide a valid resolution context.');
 
-        if (!resolutionContext.originContainer)
-            throw new DeliveryError('The provided resolution context must have a valid origin container.');
-
         if (!dependencyMetadata)
             throw new DeliveryError('Must provide the depencency metadata to deliver from.');
 
@@ -37,16 +37,10 @@ export class ContainerizedDeliveryStrategy implements IDeliveryStrategy {
         if (!dependencyMetadata.servicingStrategy)
             throw new DeliveryError('The provided dependency metadata must have a valid servicing strategy.')
 
-        let map = this._containerInstanceMap.find(map => map.container === resolutionContext.originContainer);
-        let servingResult:any;
-        
-        if (!map) {
-            servingResult = dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);
-            this._containerInstanceMap.push({ container: resolutionContext.originContainer, instance: servingResult });
-        } else
-            servingResult = map.instance;
-
-
-        return servingResult;        
+        if (!this._isServed) {
+            this._instance = dependencyMetadata.servicingStrategy.serve(resolutionContext, dependencyMetadata.activationReference);
+            this._isServed = true;
+        }
+        return this._instance;        
     }
 }
