+++ source/fluent-syntax/bindSyntax.ts
