+++ source/fluent-syntax/containerizedResolutionSyntax.ts
