+++ source/fluent-syntax/iContainerizedResolutionSyntax.ts
